+++ simplyprint_ws_client/client/config/json.py
@@ -3,9 +3,9 @@ import logging
 from pathlib import Path
 from typing import Optional
 
-from simplyprint_ws_client.config.config import Config
+from simplyprint_ws_client.client.config.config import Config
 from .manager import ConfigManager
-from ..helpers.file_backup import FileBackup
+from simplyprint_ws_client.helpers.file_backup import FileBackup
 
 
 class JsonConfigManager(ConfigManager):
