+++ simplyprint_ws_client/client/logging/__init__.py
