+++ simplyprint_ws_client/client/logging/client_logger.py
