+++ simplyprint_ws_client/client/state/__init__.py
@@ -1,2 +1,3 @@
 from .always import Always
 from .state import DEFAULT_EVENT, ClientState, State, to_event
+from .printer import *
