+++ simplyprint_ws_client/client/state/printer.py
@@ -10,11 +10,11 @@ from traitlets import Union as TraitletsUnion
 
 from .always import Always
 from .state import State, ClientState, to_event
-from ..events.client_events import CpuInfoEvent, FileProgressEvent, MachineDataEvent, FirmwareEvent, \
+from simplyprint_ws_client.events.client_events import CpuInfoEvent, FileProgressEvent, MachineDataEvent, FirmwareEvent, \
     FirmwareWarningEvent, FilamentSensorEvent, PowerControllerEvent, JobInfoEvent, LatencyEvent, WebcamStatusEvent, \
     WebcamEvent, MaterialDataEvent, StateChangeEvent, ConnectionEvent, ToolEvent
-from ..helpers.ambient_check import AmbientTemperatureState
-from ..helpers.temperature import Temperature
+from simplyprint_ws_client.helpers.ambient_check import AmbientTemperatureState
+from simplyprint_ws_client.helpers.temperature import Temperature
 
 
 class PrinterCpuFlag(Enum):
