+++ simplyprint_ws_client/connection/connection.py
@@ -8,11 +8,11 @@ from aiohttp import (ClientConnectorError, ClientSession,
                      ClientWebSocketResponse, WSMsgType,
                      WSServerHandshakeError, ClientOSError)
 
-from .client import Client
-from .events import DemandEvent, ServerEvent, EventFactory
-from .events.client_events import ClientEvent, ClientEventMode
-from .events.event import Event
-from .events.event_bus import EventBus
+from simplyprint_ws_client.client import Client
+from simplyprint_ws_client.events import DemandEvent, ServerEvent, EventFactory
+from simplyprint_ws_client.events.client_events import ClientEvent, ClientEventMode
+from simplyprint_ws_client.events.event import Event
+from simplyprint_ws_client.events.event_bus import EventBus
 
 
 class ConnectionPollEvent(Event):
