+++ simplyprint_ws_client/examples/__main__.py
@@ -1,12 +1,12 @@
 import logging
 import time
 
-from simplyprint_ws_client.app import ClientApp, ClientOptions, ClientMode
+from simplyprint_ws_client.client.app import ClientApp, ClientOptions, ClientMode
 from simplyprint_ws_client.cli import ClientCli
-from simplyprint_ws_client.config import ConfigManagerType
+from simplyprint_ws_client.client.config import ConfigManagerType
 from simplyprint_ws_client.const import SimplyPrintBackend
 from simplyprint_ws_client.examples.virtual_client import VirtualClient, VirtualConfig
-from simplyprint_ws_client.logging import ClientHandler
+from simplyprint_ws_client.client.logging import ClientHandler
 
 
 def start_client(app: ClientApp):
